backend/requirements.txt
fastapi
uvicorn
groq
google-play-scraper
sentence-transformers
scikit-learn
pandas
python-dotenv
langchain
langchain-groq
langchain-huggingface
